@@ -10,8 +10,10 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=10"
+		"node": ">=12"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -31,8 +33,8 @@
 		"fetch"
 	],
 	"devDependencies": {
-		"ava": "^3.14.0",
+		"ava": "^3.15.0",
 		"tsd": "^0.14.0",
-		"xo": "^0.36.1"
+		"xo": "^0.38.2"
 	}
 }
